EHTML as Templating Engine and Framework
EHTML - HTML Framework that allows you not to write JavaScript code. It  can be described as a set of custom elements that you can put on HTML page for different purposes and use cases. The main idea and goal of this library is to provide a convenient way to get rid of JavaScript code on the client side as much as it's possible for basic and routine stuff. I think it's awesome.